+++ src/Differ/MapDifferInterface.php
@@ -4,13 +4,14 @@ declare( strict_types = 1 );
 namespace Diff\Differ;
 
 /**
- * Interface for differ that can diff two maps.
+ * Differ that diffs associatively (holding the array keys into account).
+ * These differs can produce associative diff operations (changes and diffs).
  *
  * @since 3.2
  *
  * @license GPL-2.0+
  * @author Alexander Borisov < boshurik@gmail.com >
  */
-interface MapDifferInterface {
+interface MapDifferInterface extends Differ {
 
 }